refactor: 게시글 상세보기 상단 게시글 데이터 분리
